feat(web): move project management into Settings; home becomes pure overview
